Pratyay-Patel/requirements.txt
streamlit>=1.28.0
pypdf2>=3.0.1
faiss-cpu>=1.9.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0